+++ CHANGELOG.md
@@ -6,10 +6,18 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## 1.0.0 - 2018-01-08
 
+### Added
 - Initial release
 
-## 1.0.1 - 2018-01-19
+## 1.0.1 - 2018-01-08
+
+### Fixed
+- Missing config casters
+
+## 1.0.2 - 2018-01-20
 
 ### Added
 - Package skeleton refactoring
 - Readme update
+- Custom castable base model
+- Casters documentation